Fixes #24. Moves Common Person functions to PersonModel.
