+++ client/public/elements/pages/individual/rp-page-individual.js
@@ -39,7 +39,6 @@ export default class RpPageIndividual extends RpUtilsLanding {
 
     this._resetEleProps();
 
-
     this.AppStateModel.get().then(e => this._onAppStateUpdate(e));
   }
 
@@ -205,85 +204,15 @@ export default class RpPageIndividual extends RpUtilsLanding {
     return output;
   }
 
-  // All these machinations are particular to the UI, and won't be  changed in
-  // database IMO.  Joining contacts is pretty much a PPS thing.
   getIndividualTitles(){
-    let titles = [];
-    if (!this.individual) {
-      return titles;
-    }
-    if (typeof this.individual.hasContactInfo === 'object') {
-      let contactInfo = [];
-      if (Array.isArray(this.individual.hasContactInfo)) {
-        contactInfo = [...this.individual.hasContactInfo].sort((a,b)=>(a.rank?a.rank:100)-(b.rank?b.rank:100));
-      }
-      else {
-        contactInfo = [this.individual.hasContactInfo];
-      }
-      // If first contact is odr, than only use odr,
-      let identifier_match=null
-      if (contactInfo[0].identifier && contactInfo[0].identifier.match(/^odr/)) {
-        identifier_match=/^odr/
-      } else if (contactInfo[0].identifier && contactInfo[0].identifier.match(/^pps/)) {
-        identifier_match=/^pps/    // If first is pps only include pps
-      } else {                 // Otherwise include them all
-        identifier_match=null
-      }
-      // Next join every organization with a common title, where titles are organized
-      // by rank.
-      for (let c of contactInfo) {
-        if (!(c.title && (! identifier_match || (c.identifier && c.identifier.match(identifier_match)))))
-          continue;
-        if (! Array.isArray(c.title)) { c.title=[c.title] }
-        for (const t of c.title) {
-          let ind=titles.findIndex((have)=>have.title===t)
-          if (ind > -1) {
-            if (c.organization) {
-              if (! Array.isArray(c.organization))
-                c.organization=[c.organization]
-              for (const o of c.organization) {
-                if (titles[ind].orgs.findIndex((have)=>have===o) === -1)
-                  titles[ind].orgs.push(o)
-              }
-            }
-          } else {
-            let new_title={title:t,orgs:[]}
-            if (c.organization) {
-              if (Array.isArray(c.organization)) {
-                new_title.orgs.push(...c.organization)
-              } else {
-                new_title.orgs.push(c.organization)
-              }
-            }
-            titles.push(new_title);
-          }
-        }
-      }
-    }
-    // titles are objects; {title:string,orgs:["string"]}
-    return titles;
+    return this.PersonModel.getIndividualTitles(this.individual)
   }
   getHeadlineTitle() {
-    let title=""
-    let best=this.getIndividualTitles()[0];
-    if (best && best.title)
-      title=best.title;
-    if (best && best.orgs[0])
-      title+=`, ${best.orgs[0]}`
-    return title;
+    return this.PersonModel.getHeadlineTitle(this.individual)
   }
   getBestLabel() {
-    if (this.individual && this.individual.label) {
-      if (Array.isArray(this.individual.label)) {
-        // Prefer the shortest one? This prefers fname lname over lname, fname
-        return this.individual.label.sort((a,b)=> a.length - b.length)[0]
-      }
-      return this.individual.label
-    }
-
-    return "";
-
-    }
+    return this.PersonModel.getBestLabel(this.individual)
+  }
 
   getEmailAddresses(){
     if (!this.individual) {
